requirements.txt
allure-pytest==2.8.33
allure-python-commons==2.8.33
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
distlib==0.3.1
docopt==0.6.1
filelock==3.0.12
future==0.18.2
idna==2.10
iniconfig==1.1.1
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
nose==1.3.7
packaging==20.8
parso==0.7.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
print==1.3.0
prompt-toolkit==3.0.5
py==1.10.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==6.2.2
pytest-html==3.1.1
pytest-html-reporter==0.2.3
pytest-metadata==1.11.0
python-decouple==3.4
requests==2.25.1
rets==1.0.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
traitlets==4.3.3
urllib3==1.26.2
virtualenv==20.0.26
wcwidth==0.2.4
xmltodict==0.12.0
